Add cluster run command
This adds the command to spin up the cluster but have the logs exposed. The reason is that when I spin up the cluster, I always want to see the logs and not having it run in the background to debug potential issues.

I am opening this as draft to have a discussion on this first and it seems on shutdown it somehow gets stuck. Also naming wise users will aks what is the difference between up and run (one background, one not). Lets see what others think.