Add "cluster pull" command (#51)
Often it is useful to get the most recent version of the images. For this the most recent images must be pulled. I initially added it to the bootUp command directly but realised this could become annoying as sometimes starting a cluster could take much longer. Instead I made it a manual command.