Add option to notify test service after agent policy is applied (#229)
This adds a config option, `service_notify_signal`, to the system test runner to
allow the test runner to notify the service via a signal (e.g. SIGHUP) after the
policy has been applied by the Agent.

This is useful when the test service is connecting to the listening agent or streaming data
to the agent.

It helps address two issues that I encountered.
  1. While repeatedly testing the agent stays up with the old policy applied, but
      before the new policy is applied and the old ES data cleared the test service would
      connect and deliver data to the "old" agent/policy.
  2. For UDP streams there is no reliable way to know that the agent is ready and listening.

An alternative would be to start the test service container after the policy is applied, but
this could affect tests where the agent connects to the service (e.g. a prometheus exporter)
since the test might expect the service to be ready at startup.