Print summary after errors (#381)
The table with the summary of test results is printed before the errors in
case of failure. When there are many errors, this table can end up a
long way scroll up, or even be unavailable depending on the buffer size of
the terminal. This makes it difficult to identify the data stream whose tests
failed.

Print it at the end, after the errors, so it is always available just over the
prompt when tests fail.