Add support for constant_keyword "value" in field definition
For the constant_keyword field type Elasticsearch allows for the value to be specified in the mapping.
If the value is not specified in the mapping then the value is set based on the first document that is
indexed. This adds support for specifying a constant_keyword "value" in a field definition.

https://www.elastic.co/guide/en/elasticsearch//reference/7.12/keyword.html#constant-keyword-params

Package data streams will be able to explicitly declare a constant_keyword value and will not have to
include the field with every document sent. This gives packages more control over constant_keyword field
values. It can be used optimize storage for an integration since you could omit the constant_keyword
field entirely an event's _source.

Relates: https://github.com/elastic/package-spec/pull/194