Filter directory listings by target extension
This updates elastic-agent internals to filter for the expected
extensions when listing directories in various processes, instead of
processing all the files in a given directory.

Changed:
- data_streams/*/fields/: Only .yml files are loaded.
- _dev/build/docs/: Only .md files are considered.
- data_streams/*/elasticsearch/ingest_pipelines/: .yml and .json are
  loaded.

This prevents criptic errors while developing packages, such as:

> ERROR: creating fields validator for data stream failed: can't load fields for data stream (path: /Users/adrian/go/src/github.com/elastic/integrations/packages/carbonback_edr/data_stream/log): unmarshalling field body failed: yaml: control characters are not allowed

In this case this all was caused by a `.swp` file being present in the
directory as one of it's files was open in an editor.