add support for overriding configuration folder
Configuration folder path was hardcoded to
`user home folder/.elastic-package`. For testing purposes or for
development this created some issues, as any run of the binary would
run `EnsureInstalled`, eventually failing or overriding data.

This commit introduce a new environment variable,
`ELASTIC_PACKAGE_DATA_HOME`, that can be used to override the data
folder where this CLI unpack it's configuration and data files.

The suffix `DATA_HOME` is taken from XDG spec and leaves room for
splitting configuration and data in the future without introducing a
breaking change.