add install task to Makefile
go install builds the executable and move it to GOBIN. This is useful
for installing the tool but can become troublesome during development.

Having go install performed in the build task is confusing, as running
make build would create a "system wide" binary (as most devs have GOBIN
in their PATH).

To overcome this go install is moved to a dedicated make task: install.