add missing copy for tf lock file
Terraform Lock File support has been added https://github.com/elastic/package-spec/issues/269
but was not used by Terraform Deployer as the file was not copied from
the stage folder to the workspace.

cp * does not copy hidden files.

This commit adds copy for hidden files from the stage to the
workspace folder, so it's correctly picked up by Terraform in tests.