Add new subcommand to dump Agent Policies (#862)
Added the `elastic-stack dump agent-policies` command to dump
the agent policies created in Fleet. This command is intended to be an
exploratory tool to help analyze and debug these resources.

This new command offers different options to dump the available agent
policies in Fleet: dumping all of them, by name or by package.

For testing, there is a helper that starts a Kibana client for a mocked
server with some recorded responses.