Pass policy_template name to Kibana for system tests
When calling `POST /api/fleet/package_policies` for packages that
containing multiple policy_templates, the name of the policy_template
associated to the input under test should be specified.

If the policy_template is not specified then Kibana will default to
the first policy_template which may not contain the input or may
not contain the expected package level variables for the test.

This affected testing packages like AWS.

The error that this fixes is:

    Error: error running package system tests: could not complete test run:
    could not add data stream config to policy: could not add package to policy;
    API status code = 500; response body = {"statusCode":500,
    "error":"Internal Server Error","message":"Input template not found, unable
    to find input type logfile"}

Fixes #899