Avoid hardcoded links in READMEs (#938)
Added the option to add a YAML  file containing links definitions. These definitions
will be used to render "_dev/build/docs/*.md" files that use the new placeholder "url".
The path to this file can be defined using the ELASTIC_PACKAGE_LINKS_FILE_PATH
env var. By default, if that env var is not defined, it is looked at the root of the
repository a file named "links_table.yml".
Updated documentation about README to include this new "url" placeholder.