Add support for any shell for stack shellinit (#921)
`stack shellinit` command outputs shell code to setup the environment
and connect to the local stack instances created by the `stack` command.

Before this change this command was only supporting POSIX compatible
shells. After this commit the support has been extended so is possible
to specify, through a flag, the desired compatibility for it's output.