Support for system tests in input packages (#1041)
Add support for system tests for input packages as defined in elastic/package-spec#443.

Input packages contain their test files in the root of the package, under _dev. So far most test
files where under data_stream, so there where lots of assumptions for that. This change refactors
a lot of this code to remove some of these assumptions.