Add common dynamic mappings from ECS automatically (#1073)
Add dynamic mappings defined in a JSON file as part of the "elasticsearch" key
in the manifest file of the built package. All the mappings added
dynamically have been prefixed with a common string (_embedded_ecs)
to allow distinguish them from the ones defined by the developers.
To avoid issues in previous Elastic stack versions, the prefix is separated
by a dash instead of a dot. In versions < 8.6.0 the dot in the names caused
that packages were not installed. It was interpreted that the string
after the dot was an inner element of the map, causing validation
errors.