Fix search of repository root in Windows
When looking for the repository root in Windows, if the command hasn't
been called inside a git repository, the search will loop forever.

Finalization condition is based on checking al parent directories till
root, but root was checked with "/", that will never be the case in
Windows.

Changed to take into account the volume name and the OS-specific path
separator.