Initialize default reports directory only when needed (#1294)
`report` subcommand had some code to initialize its defaults for the
output directory. This code was executed during the initialization of
the subcommand, that is allways executed.

When running any command out of a package directory, this initialization
fails, as it depends on the package directory, and it produces an error
message.

Move this initialization to the code that gets the parameter in the
`report` subcommand, so it is only executed there, and only if the user
has not provided a value for the flag.