Explicitly install packages on system tests (#1281)
This will allow to earlier detect if one is not testing the version they
currently have in the working copy.

From 8.7.0 version, it builds and install the zip package. On older
versions it installs whatever is in the configured package registry. In
that case, if the version of the package is not available in the
repository, test execution will fail.