Load stack configuration from profile (#1428)
Commands that require an stack to connect with will try to connect with the
stack of the current profile by default. This behaviour can be overridden
by setting the usual environment variables provided by `elastic-package shellinit`.

With this change `elastic-package shellinit` is not necessary when working
with stacks managed by elastic-package. shellinit is still useful to export
the configuration for other commands or scripts.
The environment variables can also be used to work with stacks not managed
by elastic-package.