Add serverless provider (#1374)
Add support to create serverless projects and test packages using
elastic-package with a new provider (--provider serverless).
Currently, supported serverless projects in elastic-package are
observability and security. For testing packages in serverless, geoIP
fields in documents are skipped when comparing with the expected
documents, since it could not be set a specific GeoIP database to make
the tests stable.