Fix tests around pipelines with custom dataset (#1437)
The dataset value in the manifest.yml file can be overwritten to allow dataset names which are not `{package}.{dirName}`. One example of this can be found here: https://github.com/elastic/integrations/pull/7670 This is already in use today and is used by some packages. Fleet installs all the assets as expected.

When working on https://github.com/elastic/integrations/pull/7670, CI is failing. The reason is that the pipeline check did not take the custom dataset config into account. The reason why this didn't show up earlier is because the check for the template already has a similar condition in place. Likely all previous packages that used this feature did only have templates and no pipelines.