Fix validations based on fields imported from ECS (#1452)
We inject fields from ECS on different scenarios: when generating documentation,
when building packages, or when validating documents in tests.

On #1335 we did a refactor around this, to fix some validation issues, in a way that
more code is reused between all these uses. After this change, validation used field
definitions that include the already resolved external fields, but we weren't including
there information that was used by validators, what included validation of expected
and allowed values.

So these validations haven't been executed since then.

Tests have been included to try to avoid regressions related to this in the future.