Avoid encoding HTML in encoded JSON documents
JSON encoders in Go standard library encode HTML special characters by
default, so the resulting JSON is safe to use in HTML documents.

In sample documents and test results in packages we want the raw JSON,
so we can see the actual content that would be ingested when using the
integration.