Add support to filter errors in packages based on codes (#1467)
Add support to filter errores in packages based on the codes
available in package-spec. These codes are defined in a new file
"validation.yml" that should be present at the root of the package.
A new test package has been added with this new file to check
that errors are filtered.