Skip by default compareResults in pipeline tests for serverless (#1521)
This PR skips the compare results function for serverless by default. This function
is in charge of comparing the documents received by elasticsearch with the samples.
The issue is just in Serverless, since there are fields related to Geo IP database that
cannot be checked properly and causing differences in JSONs.