Record with go-vcr instead of doing our custom recording (#1566)
We record requests in some tests so we can reproduce them without having running services.

We use our own framework now, that has some limitations. It can only record GET requests, and
it can only record one response for any request, so it cannot be used for tests where the state changes.

Give a try to https://pkg.go.dev/gopkg.in/dnaeon/go-vcr.v3 for this purpose.

With go-vcr, each session is stored in a file, called "cassette", that stores the sequence of interactions. By
default it matches requests per method and per URL, and each interaction is only matched once, so multiple
requests to the same endpoint can be recorded with different responses. This seems to be enough for
all the cases we have now.