Refactors for safer handling of strings (#1578)
`strings.Cut()` was introduced in Go 1.18 for the common use case of splitting a string
in two parts. In these cases it allows to directly assign each one of the parts to variables
with more meaningful names, avoiding the risk of accessing indexes that may not exist,
and improving readability.

Leverage this function in some applicable uses cases we have here.