Fix runner shutdown, improve error message
There are two fixes in this PR. The first one is around the runner being stuck in case of errors. Because a channel is used for the errors, in case 2 errors happen in a row before the error is consumed, the go routine gets stuck. As the effect of an error is stopping the generation anyways, now on error the progress on function is aborted with return. As the errors are mostly used for logging, it raises the question if the channels are needed in the first place but further cleanup is out of scope of this PR.

The second change is around error logging for scenarios where the template generates invalid JSON. For easier debugging, it now prints out the full invalid JSON event as a log event and gives an indication on where the problem lies.