[Refactor] Refactor common code between all benchmark commands
There is quite a bit of duplicated code across all the benchmark commands. This is a first quick take on bringing it together.

Initially I wanted to only refactor 1-2 functions but one lead to the next things. I am pushing this now to get early review from others, this is not complete / not fully tested. Also potentially now that the refactoring is bigger, likeyl the best place of the common files has become a different place.