List packages with benchmarks
For someone to run benchmarks or ingest sample data it is currently not easy to find the packages that have templates for benchmarks available. This is a quick POC on how the list of integrations with benchmarks can be generated. Some ideas on how this could be used:

* Add the list to the docs page around streaming data. Downside is that it needs to be manually (or automatically) updated from time to time
* Add some command to elastic-package to list the packages with templates
* Other ideas?

The current output of the script looks as following:

```
python list_benchmark_packages.py

Package: nginx
* stubstatus-benchmark
* error-benchmark

Package: mysql
* performance-benchmark

Package: aws
* sqs-benchmark
* ec2metrics-benchmark
* billing-benchmark
* ec2logs-benchmark

Package: kubernetes
* container-benchmark
* pod-benchmark