Add progress flag if docker compose is used (#1628)
Add a new environment variable to decrease the verbosity of the
docker-compose output. It relies on the --progress flag of the 
docker-compose CLI (v2).
As port of these changes a new section has been added in the README
to list all the environment variables that elastic-package CLI takes into
account.