Add docker-compose stop before retrieving logs from service deployer (#1691)
Execute docker compose stop command before retrieving logs from
the service. This will ensure that all the service logs are retrieved and
written to the target file. For instance, this ensures that output of
commands executed after receiving SIGTERM or SIGINT signals is also
written to the target, as it happens in the script executed in the Terraform
container.