Connect services to the agent network
When independent agents are set, service containers will connect to the
network of the agent. That allows to set the same "elastic-agent" alias
that it was also set in the Elastic Agent running in the stack. As
service will not be connected to the Stack network there should not be
two containers with the same alias elastic-agent.