Remove created agents in service deployer (#1771)
Enroll agents created for custom and k8s agents directly using
the test Agent Policy created in each test, and then when the 
test finishes, elastic-package removes those created agents to 
ensure that the state is the same as at the beginning. System tests
that are using the actual Elastic Agent from the Elastic stack (created by
elastic-package stack up) keep the same behavior as before.