Use declarative definitions for asset test runner expected state (#1756)
With this architecture, each runner would have a resources method, that
generates the expected resources based on some given parameters. On
this case the only parameter would be whether the package must be
installed or not. Then test runners would apply these resources before
running the tests or on tear down stages.