Revert use of PTY for docker compose invocation in Windows (#1885)
In #1801 we introduced a virtual tty to capture docker compose output as
it is generated in normal consoles. The library we use for that doesn't work
on Windows, breaking stack management and tests depending on it.

While we find an alternative, remove the use of PTY in Windows, effectively
reverting #1801 on this platform.

Add also a test to check that elastic-package stack keeps working on Windows.