Move installation package in policy and system tests (#1892)
This PR updates the runner to include two new methods in the interface
(SetupRunner and TearDownRunner). These two methods are in charge
of creating (or deleting) the global resources required to run all the tests
for every data streams and configuration file defined in the package.
Updated runners for system and policy tests.