Move getting all tests to run inside each runner (#1895)
Move the logic to obtain all the tests to be executed by one package
into the runners. Every type of test defines a TestRunner and a Tester.
The TestRunner is in charge of returninga ll the tests to be performed
(via Tester instances). And the testers have all the logic related to perform
each test. Each one of these Tester instances must have all the information
to run its own test.