Fix race condition reading from PTTY (#1906)
Execution of fast docker compose commands are getting stuck sometimes. One
possible theory is that we are closing the pseudo-tty before the copy has finished,
and it hangs there.

Close the pseudo-TTY after waiting for the finalization of the goroutine, so we avoid
problems caused by closing it while it is being read.