Run system tests in parallel (#1909)
Adapt the code to be able to run the tests in parallel. Currently,
it is just supported running in parallel tests for the system test 
runner. Added the required configuration files in some of the 
test packages to be able to test this parallelization. Moreover, it has
been tried to decrease the log verbosity in some of the most repeated
messages about container status and agent data.