Add support for multifields added by ecs@mappings (#2035)
`ecs@mappings` component template adds multifields that are not
defined in ECS or in packages using these mappings. These fields
are present in the ingested documents and cannot be validated by
elastic-package.

Add the definitions for these multifields when `ecs@mappings` is
used and no other definition is present.

Also, improve the error reported when there is no definition for a
possible multifield.