Accept as numbers strings that can be parsed as numbers (#2065)
Similar to `numeric_keyword_fields`, add a new `string_number_fields` setting to
pipeline and system tests. With this setting, validation of fields can accept strings
as numbers if they can be parsed as float. This mimics the behaviour of Elasticsearch,
that is able to accept strings as numbers if they can be parsed like them.

Refactor `numeric_keyword_fields`, so it is applied when validating values, instead of
modifying the original value.