Fix matching of leaf fields for objects (#2054)
Validation of objects was happening only depending of the format. If there were
no subobjects, the validation was happening, and if there were objects with
subfields, the validation was not happening for the subfields. This leads to
different results with different mapping settings for the same data. For example
depending if synthetics source or `subobjects: false` is used.

Additionally, for subfields that should match the definition of the parent object,
there were cases where this match was not happening, so the test could
unexpectedly fail.

Even if usually it is recommended to add explicit wildcards in dynamic mappings,
definitions like the following one are expected to match subfields, but currently they
aren't.

```
    - name: user
      type: group
      fields:
        - name: names
          type: object
          object_type: keyword
          description: Array of user identification documents.
```

The definition above would not match for `user.names.db`.