Relax validation of subobject leaves for flattened fields (#2090)
We added some additional validations for subobjects and arrays of objects
in #1498, #1489 and related PRs. These validations only apply to packages
with spec starting on 3.0.1.

These validations rely on the structure of documents. With the adoption of
features like `subobjects: false` or synthetic source the structure is lost, and
exceptions based on spec version are not working, so the tests fail for cases
where they should not for versions of the spec older than 3.0.1. This happens
for example in the `dns` data stream of the `network_traffic` package when
LogsDB is enabled.