Review tear down test system handlers (#2252)
When using independent Elastic Agents, those agents created are
going to be removed/uninstalled at the end of the test run, so there
is no need to run those two handlers. There is a special case where
it needs to be kept the reassignment of the policy when using the
"--no-provision" flag in system tests.