Ignore environment variables in tests for existing recordings (#2309)
When creating new recordings for testing, tests that already have recordings fail if the host
doesn't match. This happens when recording from environments that don't use the usual
local values, as for example in the test case added for serverless in #2307. This can even
interfere with any other environment variable set by elastic-package stack shellinit.

With this change environment variables are only used to initialize the client when creating
new recording files and not in subsequent executions.