New stack provider for environment variables (#2298)
Add a new stack provider that can be used when elastic-package is
configured with environment variables. This provider fills the gaps
for using elastic-package for testing on clusters that are not originally
managed by elastic-package.

To use it, at least ELASTIC_PACKAGE_KIBANA_HOST and
ELASTIC_PACKAGE_ELASTICSEARCH_HOST environment variables
need to be set, other variables can be used to configure authentication,
SSL and so on.

Then, run `elastic-package stack up --provider environment`.

This provider manages:
- Local elastic-agent and initial agent policy.
- Optional local logstash and logstash output.
- If no Fleet Server is available, it configures the host, a Fleet Server policy,
  a new service token and starts a local Fleet Server.

Most commands should work if the provider is able to setup everything.