Add a new validation to use or not upload API to install packages (#2511)
This PR checks whether or not the Elastic stack allows to upload
packages via API. If it cannot be used the upload API, packages
will be installed from the Package Registry.
Until Elastic stack 8.8.2 the license required was Enterprise to use
that API endpoint. In Elastic stack 8.8.2 onwards, the license required
for this feature was changed to basic.