Update new package README template
Update the README templates and generator used in new package creation
to match the new standardized README structure. The updated template
should be better at helping users write the readme, by explaining what
is needed in each section of the template.

This also makes a few other small changes which will make it easier for
new users to create integration packages. It moves the generated readme
to `_dev/build/docs`, adds a "Do not modify" message to the README
generated during `build`, and adds a default sample-event (which is
needed to build the update readme template).