Update new package README template (#2707)
* Update new package README template

Update the README templates and generator used in new package creation
to match the new standardized README structure. The updated template
should be better at helping users write the readme, by explaining what
is needed in each section of the template.

This also makes a few other small changes which will make it easier for
new users to create integration packages. It moves the generated readme
to `_dev/build/docs`, adds a "Do not modify" message to the README
generated during `build`, and adds a default sample-event (which is
needed to build the update readme template).

* Only create new files in integration packages

* Fix whitespace

* Use new template for the test apache package readme

Rewrite the readme for the test apache package to use the new docs
format.

NOTE: This format isn't used in the published apache integration.

* Updates for new doc template format

* Add comments on adding datastream fields and events to doc template
* Generate fields in "apache" test package docs

* automatically prepend generated info to generated files; generate for all package types

* fixing tests

* formatting

* only generate to the _dev folder if not a content type, as they have validation rules against that location

* generated notice to test integrations

* missed some generated notices

* last one hopefully

* remove header function from template; fixing test and tweaking the template for build

* updating docs

* updating templated based on impl feedback from ti_abusech

---------

Co-authored-by: Kylie Meli <kylie.geller@elastic.co>