Use Bubbletea TUI library (#2923)
This replaces the TUI library used in elastic-package with "github.com/charmbracelet/bubbletea".

The previously used library has more limited functionality, and is also currently unmaintained. Switching to bubbletea will support some upcoming improvements to elastic-package, which will involve more user interaction, and will benefit from the improved UI elements that are available with bubbletea. This adds a tui package which contains the TUI components, and converts the existing users of survey to use the tui package.