internal/stack: do not remove parent of log dump directory
The previous code removed the output directory and then put a new logs directory
in it. This makes it impossible to compose a directory at the output level.

The two options to fix this were to not create the "logs" subdirectory or to
only remove the "logs" subdirectory if it exists. The latter is chosen to avoid
breaking behaviour.