Set pipeline name automatically in transform definitions (#2973)
This PR adds support to set automatically the pipeline name used in
transforms adding the current package version as a prefix.
In order to achieve that elasticsearch/transform/*/transform.yml is
managed by elastic-package as a template file (Golang text template).
And it is allowed to use a function in the template named "ingestPipelineName"
that adds the current version of the package to the pipeline.