Allow to export and edit dashboards from Serverless environments (#3007)
This PR allows to export and edit dashboards in Serverless environments.
A new UX is introduced (starting in Elastic stacks >= 9.0.0) to list just
the dashboards of a given package, if the dashboard is part of a package.
In the case, the dashboard is a new one and it still does not belong to
any package, there is an option to list all dashboards as before.