stack status cmd - add columns for image build date and VCS reference (#2057)
When dealing with snapshot builds that are constantly updated, it's difficult
to know what build of a service is really running. When viewing historic build
logs from CI this information is important to know.

This adds information about the service's image to the output of
`elastic-package stack status`. The new columns in the output are IMAGE BUILD
DATE and VCS REF. The data is read from http://label-schema.org/rc1/ labels. It
uses the org.label-schema.build-date and org.label-schema.vcs-ref labels.