[System benchmark] Start service in SetUp process to update placeholders in config (#3138)
In order to be able to update the placeholders like Hostname and
Port with the expected values in system benchmarks, the service
must be started in the SetUp method so it can update the variables
in the configuration using those placeholders accordingly.
This PR ensures that the service is setup when the configuration is
read and therefore policy variables are also set.